+++ mailscanner/rep_mrtg_style.php
@@ -117,7 +117,7 @@ if (is_writable(CACHE_DIR)) {
 
     format_report_volume($data_total_size, $size_info);
 
-    $graph = new Graph(750, 350, 0, false);
+    $graph = new Graph(850, 350, 0, false);
     //$graph->SetShadow();
     $graph->SetScale("textlin");
     //$graph->SetY2Scale("lin");